kconfig: remove PICOLIBC_MODULE, introduce PICOLIBC_USE_MODULE
Remove PICOLIBC_MODULE as it is not used.
Introduce PICOLIBC_USE_MODULE as a hidden symbol. PICOLIBC_USE_MODULE
is the actual Kconfig symbol used in the CMake code to decide if
Picolibc should be build.

The symbol is promptless to allow external users of the module, such as
Zephyr itself, to setup rules / dependencies as to when Picolibc should
be built as a module.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>