zephyr: adopt to Zephyr c_library linker property
Adopt Picolibc C library linking to use Zephyr's c_library linker target
property.

The current use of `zephyr_link_libraries(c)` are vulnerable to both
symbol deduplication and the risk that the C library may not be linked
as the final libraries.

For example if an application creates a regular CMake library like this:
  find_package(Zephyr)
  add_library(foo STATIC bar.c)
  target_link_libraries(app foo)

then `foo` does not (and should not) know about which C library to link
against. `foo` should just expect a C library to be available.

Unfortunately, the existing code of
  zephyr_link_libraries(c)

can result in a final link invocation like this:
  <linker> -lapp -lc -lfoo

because `foo` has not explicitly been linked with `c` (only app has).

Zephyr has improved its C library linking by specifying the CMake linker
rule. To support late C library definition, then Zephyr provides a
c_library linker property.

This allows Picolibc to set the c_library property which again ensures
correct linking of Picolibc when built as a module.
In the above example, the following link order is guaranteed:
  <linker> -lapp -lfoo -lc

even when `foo` is not explicitly linking to `c`.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>