machine: Eliminate setjmp-dj.h (used for i386)
setjmp-dj.h declared jmp_buf as a structure of 9 values, exactly the
same as _JBLEN for the non-__iamcu__ case. In addition, the setjmp and
longjmp implementations only checked for __iamcu__, not __unix__ or
__rtems__. Eliminating the unix/rtems condition eliminated all use of
setjmp-dj.h, so we can remove that file.

That file also had a duplicate definition of longjmp which was
missing the _Noreturn attribute, which would cause problems with C++
as the two definitions conflicted.

Signed-off-by: Keith Packard <keithp@keithp.com>